Fix: 그룹 페이지 네이게이션 위치 수정
