+++ src/pages/Layout/MainLayout/GroupSection/GroupList/index.tsx
@@ -1,14 +1,18 @@
+import { useNavigate } from 'react-router-dom'
+
 import { Flex } from '@chakra-ui/react'
 
 import { useGroupPaging } from '@/api/services/group/group.api'
 import { ActiveBrownBox } from '@/components/ActiveBrownBox'
-import { AvatarLabelWithNavigate } from '@/components/AvatarLabelWithNavigate'
+import { AvatarLabel } from '@/components/AvatarLabel'
 import { IntersectionObserverLoader } from '@/components/IntersectionObserverLoader'
 import { DATA_ERROR_MESSAGES } from '@/constants/error-message'
 import { useMemberTypeStore } from '@/stores/member-type'
 import { useSelectedGroupStore } from '@/stores/selected-group'
 
 export const GroupList = () => {
+  const navigate = useNavigate()
+
   const { data, hasNextPage, isFetchingNextPage, fetchNextPage } =
     useGroupPaging({ size: 8 })
 
@@ -33,12 +37,12 @@ export const GroupList = () => {
           onClick={() => {
             setSeletedGroup(group)
             setMemberType('GROUP')
+            navigate(`/group/${group.groupId}`)
           }}
         >
-          <AvatarLabelWithNavigate
+          <AvatarLabel
             avatarSrc={group.groupdImageUrl}
             label={group.groupName}
-            linkTo={`/group/${group.groupId}`}
           />
         </ActiveBrownBox>
       ))}
