+++ src/pages/Layout/MainLayout/MemberSection/GroupMemberSection/index.tsx
@@ -1,9 +1,10 @@
 import { BiSolidGroup } from 'react-icons/bi'
+import { useNavigate } from 'react-router-dom'
 
 import { Box, Flex } from '@chakra-ui/react'
 
 import { useGrupMemberPagingSuspense } from '@/api/services/group/member.api'
-import { AvatarLabelWithNavigate } from '@/components/AvatarLabelWithNavigate'
+import { AvatarLabel } from '@/components/AvatarLabel'
 import { PageLayout } from '@/components/PageLayout'
 import { DATA_ERROR_MESSAGES } from '@/constants/error-message'
 import { useMembersLengthStore } from '@/stores/members-length'
@@ -13,6 +14,8 @@ interface GroupMemberSectionProps {
 }
 
 export const GroupMemberSection = ({ groupId }: GroupMemberSectionProps) => {
+  const navigate = useNavigate()
+
   const { data } = useGrupMemberPagingSuspense({ groupId })
   const members = data?.pages.flatMap((page) => page.records)
   const membersLength = (data?.pages[0]?.records?.length ?? 0) - 1
@@ -31,15 +34,15 @@ export const GroupMemberSection = ({ groupId }: GroupMemberSectionProps) => {
         {members.map((member) => (
           <Box
             key={member.userId}
-            paddingY={1}
+            paddingY={1.5}
             paddingX={2}
             width="full"
             _hover={{ cursor: 'pointer', background: 'brown.50' }}
+            onClick={() => navigate(`/mypage/${member.userId}`)}
           >
-            <AvatarLabelWithNavigate
+            <AvatarLabel
               avatarSrc={member.memberImageUrl}
               label={member.userName}
-              linkTo={`/mypage/${member.userId}`}
             />
           </Box>
         ))}
